requirements.txt
flask==1.1.2
Pillow== 8.0.1
boto3==1.17.26